refactor: LottoConstants 패키지 이동
